Drop python<3.10
While it is possible that there are systems without access to modern
python out there, it's highly unlikely that they will be able to deal
with the new dependencies. We are going to switch to recommending
using containerized (docker) (or perhaps nix-based as alternative)
installations for most users and we can use whatever python version.
Non-containerized installations will be treated as an option for
power users who know what they are doing

Particular reason to do this now is that we'll be having some type
annotations using | (union type), introduced in 3.10. Thus dropping
compat code now so that it doesn't hang forever